+++ utils/compare/compare.c
@@ -1,9 +1,8 @@
-#include "../boolean/boolean.h"
-
-long long compare_number = 0;
+#include "compare.h"
+uint64_t compare_number = 0;
 
 void compare_init() { compare_number = 0; }
-int compare_char(char a, char b) {
+int  compare_char(char a, char b) {
     ++compare_number;
     if (a == b) {
         return 0;
